[refactor] 모집 글 작성 페이지 리팩토링
